behavior trees

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc7083/opmode/test/LimelightCalibrate.java
@@ -28,7 +28,7 @@ public class LimelightCalibrate  extends OpMode {
     @Override
     public void loop() {
 
-        double xDistance = limelight.getDistance(Limelight.TargetPosition.SUBMERSIBLE) - 4.5;
+        double xDistance = (double) limelight.getDistance(Limelight.TargetPosition.SUBMERSIBLE) - 4.5;
 
         telemetry.addData("Distance to Target:",xDistance);
         telemetry.update();
