+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc7083/opmode/test/LimelightOrientTest.java
@@ -31,7 +31,7 @@ public class LimelightOrientTest  extends OpMode {
     @Override
     public void loop() {
 
-        double xDistance = limelight.getDistance(Limelight.TargetPosition.SUBMERSIBLE) - 4.5;
+        double xDistance = (double) limelight.getDistance(Limelight.TargetPosition.SUBMERSIBLE) - 4.5;
         //double drivePower = limelight.positionBot();
 
         telemetry.addData("Distance to Target:",xDistance);
